class_map/requirements.txt
# --- Web Backend ---
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3

# --- Core AI / LLM Libraries ---
chromadb>=0.5.11
openai>=1.52.0
sentence-transformers>=3.0.1

# --- Search & Machine Learning ---
numpy>=1.26.4
pandas>=2.2.2
rank-bm25>=0.2.2
scikit-learn>=1.3.0

# --- General Utilities ---
python-dateutil>=2.8.2
python-dotenv>=1.0.1
requests>=2.31.0
tqdm>=4.66.0